Updated to the mentor's comment's
I removed the extra files, because I accidentally had 2 folders with 2
different projects and I just didn't use 1. Then I edited the files a
little bit for the correct display of information and the correct
operation of the application itself. And as the mentor asked, in the
README file I added logins and passwords to log in under admin and user.